Update streamlit_app.py
We want pandas to read our CSV file from that S3 bucket so we use a pandas function called read_csv  to pull the data into a dataframe we'll call my_fruit_list. 
After pulling the data into a pandas dataframe called my_fruit_list, we will ask the streamlit library to display it on the page.